+++ src/google_music_scripts/commands/upload.py
@@ -65,6 +65,12 @@ from google_music_scripts.core import get_local_songs, upload_songs
 	help="Set maximum depth of recursion when scanning for local files.\nDefault is infinite recursion."
 )
 @click.option(
+	'--no-sample',
+	is_flag=True,
+	default=False,
+	help="Don't create audio sample with ffmpeg/avconv; send empty audio sample."
+)
+@click.option(
 	'--delete-on-success',
 	is_flag=True,
 	default=False,
@@ -97,6 +103,7 @@ def upload(
 	uploader_id,
 	no_recursion,
 	max_depth,
+	no_sample,
 	delete_on_success,
 	filters,
 	album_art,
@@ -133,6 +140,7 @@ def upload(
 			mm,
 			to_upload,
 			album_art=album_art,
+			no_sample=no_sample,
 			delete_on_success=delete_on_success
 		)